refactor: DashBoardHeader 분리
